+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 <!-- markdownlint-disable first-line-h1 -->
 ## 3.10.0
 
-*2026-01-23*
+*2026-01-24*
 
 **Added**
 
@@ -17,6 +17,7 @@
 - Non-width image keywords starting with `$1` (e.g., `$1页` in Chinese)
 - [Autocompletion](./README.md#autocompletion) for the mixed MediaWiki-HTML mode
 - Missing styles for the [signatureHelp](./README.md#signaturehelp) tooltip in the MediaWiki mode
+- Highlighting of parser functions and parser function argument names in the MediaWiki mode
 
 **Changed**